[TE] add HPE Slingshot (cxi) backend

+++ mooncake-transfer-engine/include/transport/transport.h
@@ -115,14 +115,14 @@ class Transport {
         std::string peer_nic_path;
         SliceStatus status;
         TransferTask *task;
-        std::vector<uint32_t> dest_rkeys;
+        std::vector<uint64_t> dest_rkeys;
         bool from_cache;
 
         union {
             struct {
                 uint64_t dest_addr;
-                uint32_t source_lkey;
-                uint32_t dest_rkey;
+                uint64_t source_lkey;
+                uint64_t dest_rkey;
                 int lkey_index;
                 int rkey_index;
                 volatile int *qp_depth;
